🐛 Fixed strange category text rendering

+++ src/main/java/com/mrcrayfish/controllable/client/gui/ButtonBindingList.java
@@ -109,8 +109,8 @@ public class ButtonBindingList extends AbstractOptionList<ButtonBindingList.Entr
         @Override
         public void render(MatrixStack matrixStack, int x, int y, int p_230432_4_, int p_230432_5_, int itemHeight, int p_230432_7_, int p_230432_8_, boolean selected, float partialTicks)
         {
-            float labelX = ButtonBindingList.this.minecraft.currentScreen.width / 2F - this.labelWidth / 2F;
-            float labelY = y + itemHeight - 9 - 1;
+            int labelX = (int) (ButtonBindingList.this.minecraft.currentScreen.width / 2F - this.labelWidth / 2F);
+            int labelY = y + itemHeight - 9 - 1;
             ButtonBindingList.this.minecraft.fontRenderer.func_243248_b(matrixStack, this.label, labelX, labelY, 0xFFFFFFFF);
         }
     }
